refactor(cli): flatten chat command module
Move the Chat command slice to the neutral command root and update its internal export and approved test fixtures without changing public CLI, cache, or compatibility behavior.
