+++ turbo/apps/cli/src/commands/chat/send.ts
@@ -11,8 +11,8 @@ import {
 import {
   getZeroChatThreadAgentId,
   sendZeroChatEvent,
-} from "../../../lib/api/domains/zero-chat";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+} from "../../lib/api/domains/zero-chat";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 import { printChatUsageError, resolveChatThreadId } from "./shared";
 
 // Non-authoritative plain-text views of a document that carries no text part.
